backend/requirements.txt
fastapi
uvicorn[standard]
PyMuPDF
requests
python-multipart
google-genai
